Improve existing retry logic
This improves upon the existing retry logic implemented in the package by expanding the retry
conditions, allowing configuration of the retry conditions and allowing disabling retries.

By default the Management Client will retry a request 2 times, when there is a 429 status code
or if there is an error returned (except for a select couple of errors).

The configuration of the max retries and status codes that will be retried on by providing a
`RetryOptions` struct with the required values with the `WithRetryOptions` function, and retries
can be disabled by passing the `WithNoRetries` struct.