Allow using a custom object when reading users
The new v1 interface made the underlying `Management` client private, no
longer allowing a custom API client call by users of this library.

We used to have a custom method like this that allowed us to unmarshal
additional fields on the user profile, such as `locale` or `upn`, which
are present for some other connection types.

	// ReadUser is a wrapper to use our CustomUser type, but is otherwise identical to
	// calling `manager.ReadUser(subject)`.
	func (a *Auth0UserClient) ReadUser(ctx context.Context, subject string) (u *CustomUser, err error) {
		err = a.Users.Request("GET", a.Users.URI("users", subject), &u, management.Context(ctx))
		return
	}

We can no longer do this with the v1 library, as the `Request` and `URI`
methods are now inaccessible.

While it would be possible to add the UPN and Locale struct fields, we
see many other fields come across from our WAAD connection, and I am
unable to enumerate them in every other connection type.

Thus, implement a new ReadCustom method that allows passing in a
destination object to unmarshal to. This pattern should be familiar to
anyone that has called `json.Unmarshal`. As seen in the test, you can
wrap the existing user with a custom object to add any fields your
particular business logic requires.